@@ -1,17 +1,16 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
         <IsPackable>false</IsPackable>
         <RootNamespace>JHipsterNetSampleApplication.Test</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="FakeItEasy" Version="5.1.0" />
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="2.2.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-        <PackageReference Include="Moq" Version="4.10.1" />
+        <PackageReference Include="FakeItEasy" Version="5.2.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.0.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.3.0" />
+        <PackageReference Include="Moq" Version="4.13.0" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1">
           <PrivateAssets>all</PrivateAssets>
@@ -21,9 +20,9 @@
           <PrivateAssets>all</PrivateAssets>
           <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="FluentAssertions" Version="5.6.0" />
+        <PackageReference Include="FluentAssertions" Version="5.9.0" />
         <PackageReference Include="FluentAssertions.AspNetCore.Mvc" Version="2.1.0" />
-        <PackageReference Include="FluentAssertions.Json" Version="5.0.0" />
+        <PackageReference Include="FluentAssertions.Json" Version="5.2.0" />
     </ItemGroup>
 
     <ItemGroup>
