@@ -9,7 +9,7 @@ namespace JHipsterNetSampleApplication.Infrastructure {
 //            https://andrewlock.net/how-to-use-the-ioptions-pattern-for-configuration-in-asp-net-core-rc2/
 //            https://andrewlock.net/adding-validation-to-strongly-typed-configuration-objects-in-asp-net-core/
             @this.Configure<JHipsterSettings>(configuration.GetSection("jhipster"));
-            @this.AddScoped<JHipsterSettings, JHipsterSettings>();
+            @this.AddSingleton<JHipsterSettings, JHipsterSettings>();
 //it doesn't work
 //            @this.AddSingleton(resolver => resolver.GetRequiredService<IOptions<NhipsterSettings>>().Value);
             return @this;
