@@ -16,7 +16,7 @@ namespace JHipsterNetSampleApplication.Test.Web.Rest {
         public OperationResourceIntTest()
         {
             _factory = new NhipsterWebApplicationFactory<Startup>();
-            _client = _factory.CreateClient();
+            _client = _factory.WithMockUser().CreateClient();
 
             _applicationDatabaseContext = _factory.GetRequiredService<ApplicationDatabaseContext>();
 
